pyproject.toml
[project]
name = "hallmoot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "@val314159", email = "jmward@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "bottle>=0.13.4",
    "docopt>=0.6.2",
    "gevent-websocket>=0.10.1",
    "ollama>=0.6.1",
    "pyyaml>=6.0.3",
    "websocket-client>=1.9.0",
]

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"
